# frontend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CreateTransactionRequest" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "categoryId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "isRecurring" : {
            "type" : "boolean"
          },
          "tagIds" : {
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "type" : "array"
          },
          "txnDate" : {
            "format" : "date",
            "type" : "string"
          },
          "txnType" : {
            "enum" : [ "INCOME", "EXPENSE" ],
            "type" : "string"
          }
        },
        "required" : [ "amount", "txnDate", "txnType" ],
        "type" : "object"
      },
...
      "PageTransactionResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/TransactionResponse"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "items" : {
              "$ref" : "#/components/schemas/SortObject"
            },
            "type" : "array"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "Pageable" : {
        "properties" : {
          "page" : {
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "sort" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "PageableObject" : {
        "properties" : {
          "offset" : {
            "format" : "int64",
            "type" : "integer"
          },
          "pageNumber" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageSize" : {
            "format" : "int32",
            "type" : "integer"
          },
          "paged" : {
            "type" : "boolean"
          },
          "sort" : {
            "items" : {
              "$ref" : "#/components/schemas/SortObject"
            },
            "type" : "array"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ReceiptUrlResponse" : {
        "properties" : {
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "SortObject" : {
        "properties" : {
          "ascending" : {
            "type" : "boolean"
          },
          "direction" : {
            "type" : "string"
          },
          "ignoreCase" : {
            "type" : "boolean"
          },
          "nullHandling" : {
            "type" : "string"
          },
          "property" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "StoredReceipt" : {
        "properties" : {
          "bytes" : {
            "format" : "int64",
            "type" : "integer"
          },
          "mimeType" : {
            "type" : "string"
          },
          "relativePath" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TagRef" : {
        "properties" : {
          "color" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TransactionResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "categoryColor" : {
            "type" : "string"
          },
          "categoryId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "categoryName" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "hasReceipt" : {
            "type" : "boolean"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "ocrStatus" : {
            "enum" : [ "PENDING", "IN_PROGRESS", "SUCCESS", "FAILED", "RETRY" ],
            "type" : "string"
          },
          "ocrText" : {
            "type" : "string"
          },
          "originalAmount" : {
            "type" : "number"
          },
          "originalCurrency" : {
            "type" : "string"
          },
          "recurring" : {
            "type" : "boolean"
          },
          "tags" : {
            "items" : {
              "$ref" : "#/components/schemas/TagRef"
            },
            "type" : "array"
          },
          "txnDate" : {
            "format" : "date",
            "type" : "string"
          },
          "txnType" : {
            "enum" : [ "INCOME", "EXPENSE" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/api/v1/budget/transactions" : {
      "get" : {
        "operationId" : "list_10",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "type",
          "required" : false,
          "schema" : {
            "enum" : [ "INCOME", "EXPENSE" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "tagId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "pageable",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageTransactionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "budget-controller" ]
      },
      "post" : {
        "operationId" : "create_5",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTransactionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransactionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "budget-controller" ]
      }
    },
...
    "/api/v1/budget/transactions/{id}/receipt" : {
      "delete" : {
        "operationId" : "delete_7",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "receipt-controller" ]
      },
      "get" : {
        "operationId" : "download",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "token",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "format" : "byte",
                    "type" : "string"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "receipt-controller" ]
      },
      "post" : {
        "operationId" : "upload",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [ "file" ],
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StoredReceipt"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "receipt-controller" ]
      }
    },
    "/api/v1/budget/transactions/{id}/receipt/url" : {
      "get" : {
        "operationId" : "getSignedUrl",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReceiptUrlResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "receipt-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}